Fix saving default lang only (thanks ru)
When an article with no renditions other than the default is saved
immediately after publishing, the id is empty and causes an SQL error.
Wrapping the id in quotes helps, but is this the real cause, or should
the article/rendition ID be pushed to the global scope when publishing a
new article?