Automatically maintain table synchronisation
First pass at a monitoring system for the textpattern table such that
any time the table is altered by third party plugins, those changes are
automatically reflected within MLP's rendition tables on the next admin
page hit.

There are a few things this routine won't do:

* Add/modify indexes
* Push changes to the ID field (primarily because of the auto increment)

If these become an issue they can be looked at later, but for now it
should track most run-of-the-mill alterations such as add/remove field,
or alterations to the default value, type or NULL status of fields.

This should negate having to hook into glz_cf directly so that old
routine has been dropped in favour of this one, which is more extensible
and (with luck) more future-proof.