Use cookies instead of sessions
While _SESSION is convenient, it's technically unnecessary. The fact
that l10n is loaded first means other plugins that come along later and
do legitimately require sessions with classes may not work properly
because the __PHP_Incomplete_Class error appears if the session has
begun before the class definition. This affects things such as shopping
cart plugins (e.g. yab_shop) and cannot easily be resolved -- even with
an autoloader -- because MLP is a hack, not a plugin.

Hopefully this patch is a drop-in replacement for session variables,
with the possible exception that manually altering the URL to remove the
lang marker will always return the site to its default language instead
of retaining the previously used language and just not showing it in the
URL.  Some might argue this is an improvement.